Moved create_session function to main startrak file + Deleted sessions.py
